+++ src/main/java/Alumni/backend/module/domain/Profile/Chatroom.java
@@ -1,4 +1,4 @@
-package Alumni.backend.module.domain.Member;
+package Alumni.backend.module.domain.Profile;
 
 import lombok.Getter;
 import org.springframework.data.annotation.CreatedDate;
