+++ src/main/java/Alumni/backend/module/domain/Profile/ChatroomMember.java
@@ -1,4 +1,4 @@
-package Alumni.backend.module.domain.Member;
+package Alumni.backend.module.domain.Profile;
 
 import Alumni.backend.module.domain.registration.Member;
 import lombok.Getter;
@@ -7,17 +7,15 @@ import javax.persistence.*;
 
 @Getter
 @Entity
-public class Alarm {
+public class ChatroomMember {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "alarm_id", nullable = false)
+    @Column(name = "chatroom_member_id", nullable = false)
     private Long id;
 
-    @Column(nullable = false)
-    private String title;
-
-    @Column(nullable = false)
-    private String content;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "chatroom_id", nullable = false)
+    private Chatroom chatroom;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id", nullable = false)
